// src/app/interfaces/general.interface.ts
export interface iGeneral {
    do_no: string;
    do_type: string;
    doc_name: string;
    hbl_no: string;
    transit_hbl: string;
    filejob_no: string;
    shipment_code: string;
    service_code: string;
    shipment_type: string;
    original_hbl_required: string;
    mbl_no: string;
    posting_date: Date;
    po_no: string;
    transaction_code: string;
    agent_ref: string;
    contract_no: string;
    an1st_sendingdate: Date;
    an2nd_sendingdate: Date;
    an3rd_sendingdate: Date;
    cus_declaration_no: string;
}