// src/app/interfaces/routing.interface.ts
export interface iRouting{
    original_code: string;
    original_name: string;
    discharge_code: string;
    discharge_name: string;
    favorite_port_discharge: string;
    port_of_loading_code: string;
    port_of_loading_name: string;
    place_of_delivery_code: string;
    place_of_delivery_name: string;
    favorite_port_delivery: string;
    final_dest: string;
    cfs_terminal: string;
    vessel: string;
    voy: string;
    etd: string;
}